Split off pool.rs from sys.rs
