+++ vulkano/src/descriptor_set/fixed_size_pool.rs
@@ -11,13 +11,13 @@ use crate::buffer::BufferAccess;
 use crate::buffer::BufferViewRef;
 use crate::descriptor_set::descriptor::DescriptorDesc;
 use crate::descriptor_set::persistent::*;
-use crate::descriptor_set::DescriptorPool;
-use crate::descriptor_set::DescriptorPoolAlloc;
-use crate::descriptor_set::DescriptorPoolAllocError;
+use crate::descriptor_set::pool::DescriptorPool;
+use crate::descriptor_set::pool::DescriptorPoolAlloc;
+use crate::descriptor_set::pool::DescriptorPoolAllocError;
+use crate::descriptor_set::pool::UnsafeDescriptorPool;
 use crate::descriptor_set::DescriptorSet;
 use crate::descriptor_set::DescriptorSetDesc;
 use crate::descriptor_set::DescriptorSetLayout;
-use crate::descriptor_set::UnsafeDescriptorPool;
 use crate::descriptor_set::UnsafeDescriptorSet;
 use crate::device::Device;
 use crate::device::DeviceOwned;
