+++ vulkano/src/descriptor_set/mod.rs
@@ -82,13 +82,7 @@ pub use self::persistent::PersistentDescriptorSetBuildError;
 pub use self::persistent::PersistentDescriptorSetError;
 pub use self::std_pool::StdDescriptorPool;
 pub use self::std_pool::StdDescriptorPoolAlloc;
-pub use self::sys::DescriptorPool;
-pub use self::sys::DescriptorPoolAlloc;
-pub use self::sys::DescriptorPoolAllocError;
 pub use self::sys::DescriptorWrite;
-pub use self::sys::DescriptorsCount;
-pub use self::sys::UnsafeDescriptorPool;
-pub use self::sys::UnsafeDescriptorPoolAllocIter;
 pub use self::sys::UnsafeDescriptorSet;
 use crate::buffer::BufferAccess;
 use crate::device::DeviceOwned;
@@ -103,6 +97,7 @@ pub mod descriptor;
 pub mod fixed_size_pool;
 mod layout;
 pub mod persistent;
+pub mod pool;
 mod std_pool;
 mod sys;
