+++ vulkano/src/descriptor_set/persistent.rs
@@ -15,8 +15,8 @@ use crate::descriptor_set::descriptor::DescriptorImageDesc;
 use crate::descriptor_set::descriptor::DescriptorImageDescArray;
 use crate::descriptor_set::descriptor::DescriptorImageDescDimensions;
 use crate::descriptor_set::descriptor::DescriptorType;
-use crate::descriptor_set::DescriptorPool;
-use crate::descriptor_set::DescriptorPoolAlloc;
+use crate::descriptor_set::pool::DescriptorPool;
+use crate::descriptor_set::pool::DescriptorPoolAlloc;
 use crate::descriptor_set::DescriptorSet;
 use crate::descriptor_set::DescriptorSetDesc;
 use crate::descriptor_set::DescriptorSetLayout;
